refactor: change name of meta helper file
