+++ app/routes/case-studies.tsx
@@ -9,7 +9,7 @@ import { type MetaFunction, useLoaderData } from "@remix-run/react"
 import { type ReactNode } from "react"
 import { type RootLoader } from "~/root"
 import { invariantResponse } from "~/utils/invariant.server"
-import { getSocialMetas } from "~/utils/seo.server"
+import { getSocialMetas } from "~/utils/seo"
 import { getCaseStudiesPage } from "~/models/case-studies.server"
 import { CaseStudiesPage } from "~/ui/templates/CaseStudiesPage"
 import { getEPARegionsWithCaseStudies } from "~/models/epa-region.server"
