+++ app/routes/case-study.$caseStudySlug.tsx
@@ -17,7 +17,7 @@ import { invariantResponse } from "~/utils/invariant.server"
 import { GeneralErrorBoundary } from "~/routes/$"
 import { Show404 } from "~/ui/templates/404"
 import type { RootLoader } from "~/root"
-import { getSocialMetas } from "~/utils/seo.server"
+import { getSocialMetas } from "~/utils/seo"
 import type { SEOHandle } from "@nasa-gcn/remix-seo"
 import { Show500 } from "~/ui/templates/500"
 import { z } from "zod"
